DOCS-#3954: move test utils to separate directory
